Remove i03 specialness from general case
