+++ tests/conftest.py
@@ -1,26 +1,17 @@
 import importlib
 import os
+from types import ModuleType
 from unittest.mock import patch
 
 import pytest
 
 from conftest import mock_attributes_table
-from dodal.beamlines import i03
 from dodal.common.beamlines import beamline_parameters, beamline_utils
-from dodal.utils import collect_factories, make_all_devices
-
-
-@pytest.fixture(autouse=True)
-def i03_beamline_parameters():
-    with patch.dict(
-        "dodal.common.beamlines.beamline_parameters.BEAMLINE_PARAMETER_PATHS",
-        {"i03": "tests/test_data/test_beamline_parameters.txt"},
-    ) as params:
-        yield params
+from dodal.utils import make_all_devices
 
 
 @pytest.fixture(scope="function")
-def module_and_devices_for_beamline(request):
+def module_and_devices_for_beamline(request: pytest.FixtureRequest):
     beamline = request.param
     with patch.dict(os.environ, {"BEAMLINE": beamline}, clear=True):
         bl_mod = importlib.import_module("dodal.beamlines." + beamline)
@@ -35,21 +26,9 @@ def module_and_devices_for_beamline(request):
         del bl_mod
 
 
-def mock_beamline_module_filepaths(bl_name, bl_module):
+def mock_beamline_module_filepaths(bl_name: str, bl_module: ModuleType):
     if mock_attributes := mock_attributes_table.get(bl_name):
         [bl_module.__setattr__(attr[0], attr[1]) for attr in mock_attributes]
         beamline_parameters.BEAMLINE_PARAMETER_PATHS[bl_name] = (
             "tests/test_data/i04_beamlineParameters"
         )
-
-
-@pytest.fixture(scope="session")
-def i03_device_factories():
-    return [f for f in collect_factories(i03).values() if hasattr(f, "cache_clear")]
-
-
-@pytest.fixture(scope="function", autouse=True)
-def clear_device_factory_caches_after_every_test(i03_device_factories):
-    yield None
-    for f in i03_device_factories:
-        f.cache_clear()  # type: ignore
